Soluciona defecto 77,78, agrega uso de cache para recuperar lista de productos para la app y agrega mas tablas al nuevo proceso de Sync

+++ frmAdelantoEfectivo.vb
@@ -7,7 +7,7 @@
     Dim id_Sucursal As String
     Dim Nombre_Sucursal As String
     Dim Funciones As New Funciones
-    Public id_Adelanto As Integer = 0
+    Public id_Adelanto As Int64 = 0
 
     'Cuando carga el formulario.
     Private Sub frmAdelantoEfectivo_Load(ByVal sender As System.Object, ByVal e As System.EventArgs) Handles MyBase.Load
